+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- DAP: Defer automatic registration of nvim-dap configurations on LSP client init,
+  to improve reliability.
+
 ## [4.14.1] - 2024-03-17
 
 ### Fixed