+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.16.0] - 2024-03-24
+
+### Added
+
+- Initialization: Warn if nvim-lspconfig.rust-analyzer setup is detected.
+
 ## [4.15.0] - 2024-03-24
 
 ### Added